backend/requirements.txt
fastapi
uvicorn[standard]
cohere
qdrant-client
psycopg2-binary
beautifulsoup4
requests
python-dotenv
slowapi
SQLAlchemy
lxml
